# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "合同管理系统 API",
    "version": "1.0.0",
    "description": "\n# 合同管理系统 API 文档\n\n## 概述\n\n基于契约优先（Contract-First）架构的 RESTful API，用于管理华道业务的合同、用户和部门。\n\n## 认证方式\n\n系统采用 JWT 双 Token 认证机制：\n\n- **Access Token**: 用于 API 认证，有效期 15 分钟\n- **Refresh Token**: 用于刷新 Access Token，有效期 7 天\n\n### 获取 Token\n\n调用 `POST /api/auth/login` 获取 Access Token 和 Refresh Token。\n\n### 使用 Token\n\n在请求头中添加：\n```\nAuthorization: Bearer <access_token>\n```\n\n### 刷新 Token\n\n当 Access Token 过期时，调用 `POST /api/auth/refresh` 使用 Refresh Token 获取新的 Access Token。\n\n## 权限控制\n\n### 功能权限（菜单权限）\n\n用户只能看到和访问被授权的功能模块。\n\n### 数据权限\n\n- **普通用户 (EMPLOYEE)**: 只能查看自己创建的数据\n- **管理员 (ADMIN)**: 可以查看所有数据\n\n## 错误响应\n\n所有错误响应遵循 RFC 7807 Problem Details 规范：\n\n```json\n{\n  \"type\": \"https://example.com/errors/validation-error\",\n  \"title\": \"Validation Error\",\n  \"status\": 400,\n  \"detail\": \"One or more fields failed validation\",\n  \"instance\": \"/api/users\",\n  \"errors\": [\n    {\n      \"field\": \"email\",\n      \"message\": \"Invalid email format\"\n    }\n  ]\n}\n```\n\n## 分页\n\n列表接口支持分页，通过查询参数控制：\n\n- `page`: 页码（从 1 开始）\n- `limit`: 每页数量\n\n分页响应包含元数据：\n\n```json\n{\n  \"success\": true,\n  \"data\": [...],\n  \"meta\": {\n    \"page\": 1,\n    \"limit\": 20,\n    \"total\": 100,\n    \"totalPages\": 5\n  },\n  \"timestamp\": \"2026-06-08T12:00:00Z\"\n}\n```\n    ",
    "contact": {
      "name": "API Support",
      "email": "support@huada.com"
    },
...
    }
...
  "servers": [
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "description": "用户登录，返回访问令牌和刷新令牌",
        "tags": [
          "认证"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "number"
                        },
                        "username": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "realName": {
                          "type": "string"
                        },
                        "phone": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "role": {
                          "type": "string",
                          "enum": [
                            "ADMIN",
                            "EMPLOYEE"
                          ]
                        },
                        "departmentId": {
                          "anyOf": [
                            {
                              "type": "number"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "status": {
                          "default": "ACTIVE",
                          "type": "string",
                          "enum": [
                            "ACTIVE",
                            "INACTIVE",
                            "LOCKED"
                          ]
                        },
                        "isActive": {
                          "default": true,
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "username",
                        "email",
                        "realName",
                        "phone",
                        "role",
                        "departmentId",
                        "status",
                        "isActive",
                        "createdAt",
                        "updatedAt"
                      ],
                      "additionalProperties": false
                    },
                    "accessToken": {
                      "type": "string"
                    },
                    "refreshToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "user",
                    "accessToken",
                    "refreshToken",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "用户名或密码错误"
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "summary": "刷新令牌",
        "description": "使用刷新令牌获取新的访问令牌",
        "tags": [
          "认证"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "refreshToken": {
                    "type": "string"
                  }
                },
                "required": [
                  "refreshToken"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "刷新成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "accessToken": {
                      "type": "string"
                    },
                    "expiresIn": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "accessToken",
                    "expiresIn"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "刷新令牌无效或已过期"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "用户登出",
        "description": "用户登出，使当前令牌失效",
        "tags": [
          "认证"
        ],
        "responses": {
          "200": {
            "description": "登出成功"
          },
          "401": {
            "description": "未授权"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT 访问令牌认证"
      }
    }
  }
}